fix(AppService): Made strategies be simple utils functions
