+++ app/features/notifications/index.js
@@ -1,3 +1,3 @@
 export * from './services';
 export * from './factories';
-export * from './strategies';
+export * from './utils';
