--- app/features/notifications/strategies/index.js
@@ -1,2 +1,0 @@
-export { default as CanShowActivityNotificationStrategy } from './CanShowActivityNotificationStrategy';
-export { default as CanShowNotificationsOnScreenStrategy } from './CanShowNotificationsOnScreenStrategy';
